Update iso_writer.c
Changes for better management of write errors

+++ libs/libsacd/iso_writer.c
@@ -30,7 +30,7 @@
 
 #include "scarletbook_output.h"
 
-static size_t iso_write_frame(scarletbook_output_format_t *ft, const uint8_t *buf, size_t len)
+static int iso_write_frame(scarletbook_output_format_t *ft, const uint8_t *buf, size_t len)
 {
     size_t result = fwrite(buf, 1, len * SACD_LSN_SIZE, ft->fd);
 	if(result != len * SACD_LSN_SIZE)
@@ -38,7 +38,7 @@ static size_t iso_write_frame(scarletbook_output_format_t *ft, const uint8_t *bu
 		//LOG(lm_main, LOG_ERROR, ("iso_write_frame(): error writting in file.") );
 		return -1;               				
 	}
-	 return result;
+	 return (int)result;
 }
 
 scarletbook_format_handler_t const * iso_format_fn(void)
